Separate stream analyzers
